requirements.txt
requests==2.26.0
pygithub==1.55
bs4==0.0.1
packaging==21.0
beautifulsoup4==4.10.0
python-gitlab~=2.10

